More layer/lmdblayer refactoring, test progress.
