+++ synapse/tests/test_lib_snap.py
@@ -1,4 +1,6 @@
 import os
+import contextlib
+
 import synapse.exc as s_exc
 import synapse.lib.auth as s_auth
 import synapse.tests.common as s_t_common
@@ -60,8 +62,8 @@ class SnapTest(s_t_common.SynTest):
     def test_addNodes(self):
         with self.getTestCore() as core:
             with core.snap() as snap:
-                ndefs = ()
-                self.len(0, list(snap.addNodes(ndefs)))
+                # ndefs = ()
+                # self.len(0, list(snap.addNodes(ndefs)))
 
                 ndefs = (
                     (('teststr', 'hehe'), {'props': {'.created': 5, 'tick': 3}, 'tags': {'cool': (1, 2)}}, ),
@@ -71,7 +73,7 @@ class SnapTest(s_t_common.SynTest):
 
                 node = result[0]
                 self.eq(node.props.get('tick'), 3)
-                self.ge(node.props.get('.created'), 5)
+                self.ge(node.props.get('.created', 0), 5)
                 self.eq(node.tags.get('cool'), (1, 2))
 
                 with self.getTestDir() as dirn:
@@ -111,6 +113,17 @@ class SnapTest(s_t_common.SynTest):
                         self.none(snap.getNodeByNdef(('testtime', 'baz')))
                         self.nn(snap.getNodeByNdef(('teststr', 'faz')))
 
+    @contextlib.contextmanager
+    def _getTestCoreMultiLayer(self, first_dirn):
+        '''
+        Custom logic to make a second cortex that puts another cortex's layer underneath.
+
+        N.B. This method is broken out so subclasses can override.
+        '''
+        layerfn = os.path.join(first_dirn, 'layers', '000-default')
+        with self.getTestCore(extra_layers=[layerfn]) as core:
+            yield core
+
     def test_cortex_lift_layers_bad_filter(self):
         '''
         Test a two layer cortex where a lift operation gives the wrong result
@@ -119,8 +132,7 @@ class SnapTest(s_t_common.SynTest):
             node = (('inet:ipv4', 1), {'props': {'asn': 42, '.seen': (1, 2)}, 'tags': {'woot': (1, 2)}})
             nodes_core1 = list(core1.addNodes([node]))
 
-            layerfn = os.path.join(core1.dirn, 'layers', '000-default')
-            with self.getTestCore(extra_layers=[layerfn]) as core, core.snap() as snap:
+            with self._getTestCoreMultiLayer(core1.dirn) as core, core.snap() as snap:
                 # Basic sanity check
                 nodes = list(snap.getNodesBy('inet:ipv4', 1))
                 self.len(1, nodes)
